Add Vagrant Box for testing USB Gadgets in loopback with dummy_hcd
Not everybody has a USB Gadget capable Linux computer on hand nor is it
always desirable to develop and test directly on "target" devices.

This Vagrant box provides a reproducible development environment with
USB Gadget Loopback support provided by dummy_hcd.

Unfortunately, dummy_hcd is not compiled by default in Ubuntu/Debian,
even in their "extra" package. We must compile and install it ourselves.
This Vagrant box tries to ease that process and make it easily
reproducible.

The vagrant box also contains Rust, so code in this repository can be
built, tested, and run.